add exceptions to article, users, and make custom exception in exceptions file

+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from schemas import UserBase
 from db.models import DbUser
+from fastapi import HTTPException, status
 
 def create_user(db: Session, request: UserBase):
   new_user = DbUser(
@@ -18,12 +19,17 @@ def get_all_users(db: Session):
   return db.query(DbUser).all()
 
 def get_user(db: Session, id: int):
-  #Handle any exceptions
-  return db.query(DbUser).filter(DbUser.id == id).first()
+  user = db.query(DbUser).filter(DbUser.id == id).first()
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+      detail=f'User with id {id} not found')
+  return 
 
 def update_user(db: Session, id: int, request: UserBase):
   user = db.query(DbUser).filter(DbUser.id == id)
-  #Handle any exceptions
+  if not user.first():
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+      detail=f'User with id {id} not found')
   user.update({
     DbUser.username: request.username,
     DbUser.email: request.email,
@@ -34,7 +40,9 @@ def update_user(db: Session, id: int, request: UserBase):
 
 def delete_user(db: Session, id: int):
   user = db.query(DbUser).filter(DbUser.id == id).first()
-  #Handle any exceptions
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+      detail=f'User with id {id} not found')
   db.delete(user)
   db.commit()
   return 'ok'
